# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "族谱管理系统的RESTful API文档",
        "title": "族谱管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/relationship/createParentChild": {
            "post": {
                "description": "创建父母与子女之间的关系",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "亲子关系管理"
                ],
                "summary": "创建亲子关系",
                "parameters": [
                    {
                        "description": "亲子关系信息",
                        "name": "relation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/relationship.ParentChild"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/relationship.ParentChild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/relationship/deleteParentChild": {
            "delete": {
                "description": "删除父母与子女之间的关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "亲子关系管理"
                ],
                "summary": "删除亲子关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "父母ID",
                        "name": "parentId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "子女ID",
                        "name": "childId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/relationship/getChildrenByParentId": {
            "get": {
                "description": "根据父母ID获取其所有子女",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "亲子关系管理"
                ],
                "summary": "获取父母的子女",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "父母ID",
                        "name": "parentId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/relationship.ParentChild"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/relationship/getParentChild": {
            "get": {
                "description": "获取父母与子女之间的关系",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "亲子关系管理"
                ],
                "summary": "获取亲子关系",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "父母ID",
                        "name": "parentId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "子女ID",
                        "name": "childId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/relationship.ParentChild"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/relationship/getParentsByChildId": {
            "get": {
                "description": "根据子女ID获取其所有父母",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "亲子关系管理"
                ],
                "summary": "获取子女的父母",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "子女ID",
                        "name": "childId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/relationship.ParentChild"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "relationship.ParentChild": {
            "type": "object",
            "properties": {
                "child_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_primary": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "relation_type": {
                    "type": "string"
                }
            }
        }
    }
}